server/requirements.txt
flask
flask-cors
Pillow
PyMuPDF
PyPDF2
pdf2image

